requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tzlocal==5.2
uritemplate==4.1.1
gunicorn==23.0.0
whitenoise==6.8.2
django-storages[boto3]==1.14.4